requirements.txt
asgiref==3.5.2
cfgv==3.3.1
distlib==0.3.4
Django==4.0.5
djangorestframework==3.13.1
filelock==3.7.1
identify==2.5.1
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.19.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.15.1
